starter-project/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "litellm>=1.72.6",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
